requirements.txt
torch
numpy
pillow
matplotlib